+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 2.x branch
 ## 2.7 branch
 ### 2.7.0
-* `BackupTrait::getTarget()` and `BackupTrait::getDriverName()` have been removed.
+* `BackupTrait::getTarget()`, `BackupTrait::getDriverName()` and
+    `BackupTrait::getValidExtensions()` methods have been removed.
 
 ## 2.6 branch
 ### 2.6.6